+++ Controller/Products/Sendupdated.php
@@ -66,8 +66,9 @@ class Sendupdated extends \Autocompleteplus\Autosuggest\Controller\Products
         $to = $this->getRequest()->getParam('to', false);
         $page = $this->getRequest()->getParam('page', 1);
         $send_oos = $this->getRequest()->getParam('send_oos', false);
-
         $send_oos = boolval($send_oos);
+
+        $this->xmlGenerator->checkCachedAttrValues($storeId);
         $catalogXml = $this->xmlGenerator
             ->renderUpdatesCatalogXml($count, $storeId, $from, $to, $page, $send_oos);